LayoutTests/fast/url/segments-userinfo-vs-host-expected.txt

Canonicalization of URLs that start with something that may or may not be userinfo

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canonicalize('http:@www.apple.com') is 'http://www.apple.com/'
PASS segments('http:@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http:/@www.apple.com') is 'http://www.apple.com/'
PASS segments('http:/@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http://@www.apple.com') is 'http://www.apple.com/'
PASS segments('http://@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http:a:b@www.apple.com') is 'http://a:b@www.apple.com/'
PASS segments('http:a:b@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http:/a:b@www.apple.com') is 'http://a:b@www.apple.com/'
PASS segments('http:/a:b@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http://a:b@www.apple.com') is 'http://a:b@www.apple.com/'
PASS segments('http://a:b@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http://@pple.com') is 'http://pple.com/'
PASS segments('http://@pple.com') is '["http:","pple.com","","/","",""]'

PASS canonicalize('http::b@www.apple.com') is 'http://:b@www.apple.com/'
PASS segments('http::b@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http:/:b@www.apple.com') is 'http://:b@www.apple.com/'
PASS segments('http:/:b@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http://:b@www.apple.com') is 'http://:b@www.apple.com/'
PASS segments('http://:b@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http:/:@/www.apple.com') is 'http:/:@/www.apple.com'
PASS segments('http:/:@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http://user@/www.apple.com') is 'http://user@/www.apple.com'
PASS segments('http://user@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http:@/www.apple.com') is 'http:@/www.apple.com'
PASS segments('http:@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http:/@/www.apple.com') is 'http:/@/www.apple.com'
PASS segments('http:/@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http://@/www.apple.com') is 'http://@/www.apple.com'
PASS segments('http://@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('https:@/www.apple.com') is 'https:@/www.apple.com'
PASS segments('https:@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http:a:b@/www.apple.com') is 'http:a:b@/www.apple.com'
PASS segments('http:a:b@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http:/a:b@/www.apple.com') is 'http:/a:b@/www.apple.com'
PASS segments('http:/a:b@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http://a:b@/www.apple.com') is 'http://a:b@/www.apple.com'
PASS segments('http://a:b@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http::@/www.apple.com') is 'http::@/www.apple.com'
PASS segments('http::@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http:a:@www.apple.com') is 'http://a@www.apple.com/'
PASS segments('http:a:@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http:/a:@www.apple.com') is 'http://a@www.apple.com/'
PASS segments('http:/a:@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http://a:@www.apple.com') is 'http://a@www.apple.com/'
PASS segments('http://a:@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http://a:b@www.@pple.com/p@th') is 'http://a:b@www.@pple.com/p@th'
PASS segments('http://a:b@www.@pple.com/p@th') is '[":","","","","",""]'

PASS canonicalize('http://www.@@pple.com') is 'http://www.@@pple.com'
PASS segments('http://www.@@pple.com') is '[":","","","","",""]'

PASS canonicalize('http://@@pple.com') is 'http://@@pple.com'
PASS segments('http://@@pple.com') is '[":","","","","",""]'

PASS canonicalize('http://@@@pple.com') is 'http://@@@pple.com'
PASS segments('http://@@@pple.com') is '[":","","","","",""]'

PASS canonicalize('http:@@/www.apple.com') is 'http:@@/www.apple.com'
PASS segments('http:@@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http:/@@/www.apple.com') is 'http:/@@/www.apple.com'
PASS segments('http:/@@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http://@@/www.apple.com') is 'http://@@/www.apple.com'
PASS segments('http://@@/www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http:@:b@www.apple.com') is 'http:@:b@www.apple.com'
PASS segments('http:@:b@www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http:/@:b@www.apple.com') is 'http:/@:b@www.apple.com'
PASS segments('http:/@:b@www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http://@:b@www.apple.com') is 'http://@:b@www.apple.com'
PASS segments('http://@:b@www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http://www.@pple.com') is 'http://www.@pple.com/'
PASS segments('http://www.@pple.com') is '["http:","pple.com","","/","",""]'

PASS canonicalize('http:@:www.apple.com') is 'http:@:www.apple.com'
PASS segments('http:@:www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http:/@:www.apple.com') is 'http:/@:www.apple.com'
PASS segments('http:/@:www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http://@:www.apple.com') is 'http://@:www.apple.com'
PASS segments('http://@:www.apple.com') is '[":","","","","",""]'

PASS canonicalize('http://:@www.apple.com') is 'http://www.apple.com/'
PASS segments('http://:@www.apple.com') is '["http:","www.apple.com","","/","",""]'

PASS canonicalize('http:@:/www.apple.com') is 'http://:/www.apple.com'
PASS segments('http:@:/www.apple.com') is '["http:","","0","/www.apple.com","",""]'

PASS canonicalize('http:/@:/www.apple.com') is 'http://:/www.apple.com'
PASS segments('http:/@:/www.apple.com') is '["http:","","0","/www.apple.com","",""]'

PASS canonicalize('http://@:/www.apple.com') is 'http://:/www.apple.com'
PASS segments('http://@:/www.apple.com') is '["http:","","0","/www.apple.com","",""]'

PASS canonicalize('http:@:80/www.apple.com') is 'http://:80/www.apple.com'
PASS segments('http:@:80/www.apple.com') is '["http:","","80","/www.apple.com","",""]'

PASS successfullyParsed is true

TEST COMPLETE
